Create a Tuner GUI Interface so different UI types can be built
Still need to add the part where you can select the different
tuner types in User Settings but that can be done when the 2nd
tuner UI is created.
